Add utility to extract all MethodInstances (#6)
This is aimed at understanding whether a cache file might be missing
certain code. By collecting all MethodInstances at the time of loading
and comparing them against new inferences measured with
SnoopCompileCore when workloads are executed, we gain the ability
to check whether omission-from-cache or some other factor caused
the need for compilation.